refactor: remove unused InlineKeyboardToDbButtons from production code
- Move InlineKeyboardToDbButtons from content.go to content_test.go
- Add //go:build testtools tag to content_test.go
- Add net/url import to test file
- Function is only used in tests, not production code
